Moved and renamed mysql model to repository
