+++ src/Storage/MySQL/Repository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Projom\Storage\Model;
+namespace Projom\Storage\MySQL;
 
 use Projom\Storage\Query\MySQLQuery;
 use Projom\Storage\SQL\Util\Aggregate;
@@ -10,10 +10,10 @@ use Projom\Storage\SQL\Util\Operator;
 use Projom\Storage\Util;
 
 /**
- * MySQLModel is a trait that provides a set of methods to interact with a database table.
+ * Repository is a trait that provides a set of methods to interact with a database table.
  * 
  * How to use:
- * * Use this trait to create a query-able "model/repository" of the class using the trait.
+ * * Use this trait to create a query-able "repository" of the class using the trait.
  * * The name of the class using the trait should be the same as the database table name.
  *
  * Mandatory abstract methods to implement: 
@@ -23,9 +23,9 @@ use Projom\Storage\Util;
  * * formatFields(): array [ 'Field' => 'string', 'AnotherField' => 'int', ... ]
  * * redactFields(): array [ 'Field', 'AnotherField' ]
  * 
- * The value of all redacted fields will be replaced with the string "\_\_REDACTED\_\_".
+ * The value of all redacted fields will be replaced with the string "__REDACTED__".
  */
-trait MySQLModel
+trait Repository
 {
 	private $table = null;
 	private $primaryField = null;
